refactor karakeep exporter
