+++ src/karakeep/features/exporter/karakeep_exporter.lua
@@ -1,10 +1,9 @@
-local EventListener = require('ui/widget/eventlistener')
+local BaseExporter = require('base')
 local logger = require('logger')
 local _ = require('gettext')
 
-local Provider = require('provider')
-local Notification = require('karakeep/shared/widgets/notification')
 local KarakeepMetadata = require('karakeep/shared/karakeep_metadata')
+local Notification = require('karakeep/shared/widgets/notification')
 
 ---@class BookNotes
 ---@field title string
@@ -24,19 +23,28 @@ local KarakeepMetadata = require('karakeep/shared/karakeep_metadata')
 ---@field time number
 ---@field chapter string
 
----@class KarakeepProvider : EventListener
+---@class KarakeepExporter
 ---@field ui UI Reference to UI for accessing registered modules
-local KarakeepProvider = EventListener:extend({})
+local KarakeepExporter = {}
 
-function KarakeepProvider:init()
-    local KarakeepExporter = require('karakeep/features/exporter/modules/karakeep_exporter')
-    Provider:register('exporter', 'karakeep', KarakeepExporter)
+---Create a new exporter instance with UI dependency injection
+---@param config {ui: UI} Configuration with UI reference
+---@return table exporter_instance The configured exporter instance
+function KarakeepExporter:new(config)
+    local instance = BaseExporter:new({
+        name = 'karakeep',
+        label = 'karakeep',
+        is_remote = true,
+    })
+    setmetatable(instance, { __index = self })
+    instance.ui = config.ui
+    return instance
 end
 
 ---Create a new bookmark in Karakeep
 ---@param params {title: string, content: string} The bookmark data
 ---@return table|nil result The created bookmark data if successful
-function KarakeepProvider:createBookmark(params)
+function KarakeepExporter:createBookmark(params)
     local result, error = self.ui.karakeep_api:createNewBookmark({
         body = {
             type = 'text',
@@ -46,25 +54,25 @@ function KarakeepProvider:createBookmark(params)
     })
 
     if error then
-        logger.err('[KarakeepProvider] Failed to create bookmark:', error.message)
+        logger.err('[KarakeepExporter] Failed to create bookmark:', error.message)
         Notification:error(_('Failed to create Karakeep bookmark'))
         return nil
     end
 
     if not result or not result.id then
-        logger.err('[KarakeepProvider] Invalid response: missing bookmark ID')
+        logger.err('[KarakeepExporter] Invalid response: missing bookmark ID')
         Notification:error(_('Failed to create Karakeep bookmark'))
         return nil
     end
 
-    logger.dbg('[KarakeepProvider] Created bookmark with ID:', result.id)
+    logger.dbg('[KarakeepExporter] Created bookmark with ID:', result.id)
     return result
 end
 
 ---Update an existing bookmark in Karakeep
 ---@param params {id: string, content: string} The bookmark update data
 ---@return table|nil result The updated bookmark data if successful
-function KarakeepProvider:updateBookmark(params)
+function KarakeepExporter:updateBookmark(params)
     local result, error = self.ui.karakeep_api:updateBookmark(params.id, {
         body = {
             type = 'text',
@@ -73,27 +81,34 @@ function KarakeepProvider:updateBookmark(params)
     })
 
     if error then
-        logger.err('[KarakeepProvider] Failed to update bookmark:', error.message)
+        logger.err('[KarakeepExporter] Failed to update bookmark:', error.message)
         Notification:error(_('Failed to update Karakeep bookmark'))
         return nil
     end
 
-    logger.dbg('[KarakeepProvider] Updated bookmark:', params.id)
+    logger.dbg('[KarakeepExporter] Updated bookmark:', params.id)
     return result
 end
 
----Handle export request from KarakeepExporter
----@param books_with_markdown table[] Array of {book: BookNotes, markdown_content: string}
-function KarakeepProvider:onExportToKarakeep(books_with_markdown)
-    logger.info('[KarakeepProvider] Starting export of', #books_with_markdown, 'books')
+---Main export method called by KOReader's exporter system
+---@param book_notes BookNotes[] Array of book notes to export
+---@return boolean success True if export was successful
+function KarakeepExporter:export(book_notes)
+    logger.info('[KarakeepExporter] Starting export of', #book_notes, 'books')
 
     local success_count = 0
     local error_count = 0
 
-    for _, item in ipairs(books_with_markdown) do
-        local book = item.book
-        local markdown_content = item.markdown_content
+    for _, book in ipairs(book_notes) do
         local bookmark_data = KarakeepMetadata.getBookmark(book.file)
+
+        -- Generate markdown content inline
+        local md = require('template/md')
+        -- selene: allow(undefined_variable)
+        local plugin_settings = G_reader_settings:readSetting('exporter') or {}
+        local markdown_settings = plugin_settings.markdown or {}
+        local markdown_table = md.prepareBookContent(book, markdown_settings)
+        local markdown_content = table.concat(markdown_table, '\n')
 
         if bookmark_data and bookmark_data.id then
             local result = self:updateBookmark({
@@ -136,12 +151,13 @@ function KarakeepProvider:onExportToKarakeep(books_with_markdown)
     end
 
     logger.info(
-        '[KarakeepProvider] Export completed:',
+        '[KarakeepExporter] Export completed:',
         success_count,
         'success,',
         error_count,
         'errors'
     )
+    return error_count == 0
 end
 
-return KarakeepProvider
+return KarakeepExporter
